update with decimal stuff
