+++ table/arrow_utils.go
@@ -28,14 +28,12 @@ import (
 	"github.com/apache/arrow-go/v18/arrow"
 	"github.com/apache/arrow-go/v18/arrow/array"
 	"github.com/apache/arrow-go/v18/arrow/compute"
-	"github.com/apache/arrow-go/v18/arrow/decimal128"
 	"github.com/apache/arrow-go/v18/arrow/extensions"
 	"github.com/apache/arrow-go/v18/arrow/memory"
 	"github.com/apache/iceberg-go"
 	"github.com/apache/iceberg-go/internal"
 	iceio "github.com/apache/iceberg-go/io"
 	tblutils "github.com/apache/iceberg-go/table/internal"
-	"github.com/google/uuid"
 	"github.com/pterm/pterm"
 )
 
@@ -1046,406 +1044,6 @@ func must[T any](v T, err error) T {
 	return v
 }
 
-func primitiveToPhysicalType(typ iceberg.Type) string {
-	switch t := typ.(type) {
-	case iceberg.BooleanType:
-		return "BOOLEAN"
-	case iceberg.Int32Type:
-		return "INT32"
-	case iceberg.Int64Type:
-		return "INT64"
-	case iceberg.Float32Type:
-		return "FLOAT"
-	case iceberg.Float64Type:
-		return "DOUBLE"
-	case iceberg.DateType:
-		return "INT32"
-	case iceberg.TimeType:
-		return "INT64"
-	case iceberg.TimestampType:
-		return "INT64"
-	case iceberg.TimestampTzType:
-		return "INT64"
-	case iceberg.StringType:
-		return "BYTE_ARRAY"
-	case iceberg.UUIDType:
-		return "FIXED_LEN_BYTE_ARRAY"
-	case iceberg.FixedType:
-		return "FIXED_LEN_BYTE_ARRAY"
-	case iceberg.BinaryType:
-		return "BYTE_ARRAY"
-	case iceberg.DecimalType:
-		switch {
-		case t.Precision() <= 9:
-			return "INT32"
-		case t.Precision() <= 18:
-			return "INT64"
-		default:
-			return "FIXED_LEN_BYTE_ARRAY"
-		}
-	default:
-		panic(fmt.Errorf("expected primitive type, got: %s", typ))
-	}
-}
-
-type typedStat[T iceberg.LiteralType] interface {
-	Min() T
-	Max() T
-}
-
-type wrappedBinaryStats struct {
-	*metadata.ByteArrayStatistics
-}
-
-func (w wrappedBinaryStats) Min() []byte {
-	return w.ByteArrayStatistics.Min()
-}
-
-func (w wrappedBinaryStats) Max() []byte {
-	return w.ByteArrayStatistics.Max()
-}
-
-type wrappedStringStats struct {
-	*metadata.ByteArrayStatistics
-}
-
-func (w wrappedStringStats) Min() string {
-	data := w.ByteArrayStatistics.Min()
-
-	return unsafe.String(unsafe.SliceData(data), len(data))
-}
-
-func (w wrappedStringStats) Max() string {
-	data := w.ByteArrayStatistics.Max()
-
-	return unsafe.String(unsafe.SliceData(data), len(data))
-}
-
-type wrappedUUIDStats struct {
-	*metadata.FixedLenByteArrayStatistics
-}
-
-func (w wrappedUUIDStats) Min() uuid.UUID {
-	uid, err := uuid.FromBytes(w.FixedLenByteArrayStatistics.Min())
-	if err != nil {
-		panic(err)
-	}
-
-	return uid
-}
-
-func (w wrappedUUIDStats) Max() uuid.UUID {
-	uid, err := uuid.FromBytes(w.FixedLenByteArrayStatistics.Max())
-	if err != nil {
-		panic(err)
-	}
-
-	return uid
-}
-
-type wrappedFLBAStats struct {
-	*metadata.FixedLenByteArrayStatistics
-}
-
-func (w wrappedFLBAStats) Min() []byte {
-	return w.FixedLenByteArrayStatistics.Min()
-}
-
-func (w wrappedFLBAStats) Max() []byte {
-	return w.FixedLenByteArrayStatistics.Max()
-}
-
-type wrappedDecStats struct {
-	*metadata.FixedLenByteArrayStatistics
-	scale int
-}
-
-func (w wrappedDecStats) Min() iceberg.Decimal {
-	dec, err := tblutils.BigEndianToDecimal(w.FixedLenByteArrayStatistics.Min())
-	if err != nil {
-		panic(err)
-	}
-
-	return iceberg.Decimal{Val: dec, Scale: w.scale}
-}
-
-func (w wrappedDecStats) Max() iceberg.Decimal {
-	dec, err := tblutils.BigEndianToDecimal(w.FixedLenByteArrayStatistics.Max())
-	if err != nil {
-		panic(err)
-	}
-
-	return iceberg.Decimal{Val: dec, Scale: w.scale}
-}
-
-type statsAgg interface {
-	min() iceberg.Literal
-	max() iceberg.Literal
-	update(stats metadata.TypedStatistics)
-	minAsBytes() ([]byte, error)
-	maxAsBytes() ([]byte, error)
-}
-
-type statsAggregator[T iceberg.LiteralType] struct {
-	curMin iceberg.TypedLiteral[T]
-	curMax iceberg.TypedLiteral[T]
-
-	cmp           iceberg.Comparator[T]
-	primitiveType iceberg.PrimitiveType
-	truncLen      int
-}
-
-func newStatAgg[T iceberg.LiteralType](typ iceberg.PrimitiveType, trunc int) statsAgg {
-	var z T
-
-	return &statsAggregator[T]{
-		primitiveType: typ,
-		truncLen:      trunc,
-		cmp:           iceberg.NewLiteral(z).(iceberg.TypedLiteral[T]).Comparator(),
-	}
-}
-
-func createStatsAgg(typ iceberg.PrimitiveType, physicalTypeStr string, truncLen int) (statsAgg, error) {
-	expectedPhysical := primitiveToPhysicalType(typ)
-	if physicalTypeStr != expectedPhysical {
-		switch {
-		case physicalTypeStr == "INT32" && expectedPhysical == "INT64":
-		case physicalTypeStr == "FLOAT" && expectedPhysical == "DOUBLE":
-		default:
-			return nil, fmt.Errorf("unexpected physical type %s for %s, expected %s",
-				physicalTypeStr, typ, expectedPhysical)
-		}
-	}
-
-	switch physicalTypeStr {
-	case "BOOLEAN":
-		return newStatAgg[bool](typ, truncLen), nil
-	case "INT32":
-		switch typ.(type) {
-		case iceberg.DecimalType:
-			return &decAsIntAgg[int32]{
-				newStatAgg[int32](typ, truncLen).(*statsAggregator[int32]),
-			}, nil
-		}
-
-		return newStatAgg[int32](typ, truncLen), nil
-	case "INT64":
-		switch typ.(type) {
-		case iceberg.DecimalType:
-			return &decAsIntAgg[int64]{
-				newStatAgg[int64](typ, truncLen).(*statsAggregator[int64]),
-			}, nil
-		}
-
-		return newStatAgg[int64](typ, truncLen), nil
-	case "FLOAT":
-		return newStatAgg[float32](typ, truncLen), nil
-	case "DOUBLE":
-		return newStatAgg[float64](typ, truncLen), nil
-	case "FIXED_LEN_BYTE_ARRAY":
-		switch typ.(type) {
-		case iceberg.UUIDType:
-			return newStatAgg[uuid.UUID](typ, truncLen), nil
-		case iceberg.DecimalType:
-			return newStatAgg[iceberg.Decimal](typ, truncLen), nil
-		default:
-			return newStatAgg[[]byte](typ, truncLen), nil
-		}
-
-		return newStatAgg[[]byte](typ, truncLen), nil
-	case "BYTE_ARRAY":
-		if typ.Equals(iceberg.PrimitiveTypes.String) {
-			return newStatAgg[string](typ, truncLen), nil
-		}
-
-		return newStatAgg[[]byte](typ, truncLen), nil
-	default:
-		return nil, fmt.Errorf("unsupported physical type: %s", physicalTypeStr)
-	}
-}
-
-func (s *statsAggregator[T]) min() iceberg.Literal { return s.curMin }
-func (s *statsAggregator[T]) max() iceberg.Literal { return s.curMax }
-
-func (s *statsAggregator[T]) update(stats metadata.TypedStatistics) {
-	st := stats.(typedStat[T])
-	s.updateMin(st.Min())
-	s.updateMax(st.Max())
-}
-
-func (s *statsAggregator[T]) toBytes(val iceberg.Literal) ([]byte, error) {
-	v, err := val.To(s.primitiveType)
-	if err != nil {
-		return nil, err
-	}
-
-	return v.MarshalBinary()
-}
-
-func (s *statsAggregator[T]) updateMin(val T) {
-	if s.curMin == nil {
-		s.curMin = iceberg.NewLiteral(val).(iceberg.TypedLiteral[T])
-	} else {
-		if s.cmp(val, s.curMin.Value()) < 0 {
-			s.curMin = iceberg.NewLiteral(val).(iceberg.TypedLiteral[T])
-		}
-	}
-}
-
-func (s *statsAggregator[T]) updateMax(val T) {
-	if s.curMax == nil {
-		s.curMax = iceberg.NewLiteral(val).(iceberg.TypedLiteral[T])
-	} else {
-		if s.cmp(val, s.curMax.Value()) > 0 {
-			s.curMax = iceberg.NewLiteral(val).(iceberg.TypedLiteral[T])
-		}
-	}
-}
-
-func (s *statsAggregator[T]) minAsBytes() ([]byte, error) {
-	if s.curMin == nil {
-		return nil, nil
-	}
-
-	if s.truncLen > 0 {
-		return s.toBytes((&iceberg.TruncateTransform{Width: s.truncLen}).
-			Apply(iceberg.Optional[iceberg.Literal]{Valid: true, Val: s.curMin}).Val)
-	}
-
-	return s.toBytes(s.curMin)
-}
-
-func (s *statsAggregator[T]) maxAsBytes() ([]byte, error) {
-	if s.curMax == nil {
-		return nil, nil
-	}
-
-	if s.truncLen <= 0 {
-		return s.toBytes(s.curMax)
-	}
-
-	switch s.primitiveType.(type) {
-	case iceberg.StringType:
-		if !s.curMax.Type().Equals(s.primitiveType) {
-			return nil, errors.New("expected current max to be a string")
-		}
-
-		curMax := any(s.curMax.Value()).(string)
-		result := truncateUpperBoundText(curMax, s.truncLen)
-		if result != "" {
-			return s.toBytes(iceberg.StringLiteral(result))
-		}
-
-		return nil, nil
-	case iceberg.BinaryType:
-		if !s.curMax.Type().Equals(s.primitiveType) {
-			return nil, errors.New("expected current max to be a binary")
-		}
-
-		curMax := any(s.curMax.Value()).([]byte)
-		result := truncateUpperBoundBinary(curMax, s.truncLen)
-		if len(result) > 0 {
-			return s.toBytes(iceberg.BinaryLiteral(result))
-		}
-
-		return nil, nil
-	default:
-		return nil, fmt.Errorf("%s cannot be truncated for upper bound", s.primitiveType)
-	}
-}
-
-type decAsIntAgg[T int32 | int64] struct {
-	*statsAggregator[T]
-}
-
-func (s *decAsIntAgg[T]) minAsBytes() ([]byte, error) {
-	if s.curMin == nil {
-		return nil, nil
-	}
-
-	lit := iceberg.DecimalLiteral(iceberg.Decimal{
-		Val:   decimal128.FromI64(int64(s.curMin.Value())),
-		Scale: s.primitiveType.(iceberg.DecimalType).Scale(),
-	})
-	if s.truncLen > 0 {
-		return s.toBytes((&iceberg.TruncateTransform{Width: s.truncLen}).
-			Apply(iceberg.Optional[iceberg.Literal]{Valid: true, Val: lit}).Val)
-	}
-
-	return s.toBytes(lit)
-}
-
-func (s *decAsIntAgg[T]) maxAsBytes() ([]byte, error) {
-	if s.curMax == nil {
-		return nil, nil
-	}
-
-	lit := iceberg.DecimalLiteral(iceberg.Decimal{
-		Val:   decimal128.FromI64(int64(s.curMax.Value())),
-		Scale: s.primitiveType.(iceberg.DecimalType).Scale(),
-	})
-	if s.truncLen <= 0 {
-		return s.toBytes(lit)
-	}
-
-	return nil, fmt.Errorf("%s cannot be truncated for upper bound", s.primitiveType)
-}
-
-type metricModeType string
-
-const (
-	metricModeTruncate metricModeType = "truncate"
-	metricModeNone     metricModeType = "none"
-	metricModeCounts   metricModeType = "counts"
-	metricModeFull     metricModeType = "full"
-)
-
-type metricsMode struct {
-	typ metricModeType
-	len int
-}
-
-var truncationExpr = regexp.MustCompile(`^truncate\((\d+)\)$`)
-
-func matchMetricsMode(mode string) (metricsMode, error) {
-	sanitized := strings.ToLower(strings.TrimSpace(mode))
-	if strings.HasPrefix(sanitized, string(metricModeTruncate)) {
-		m := truncationExpr.FindStringSubmatch(sanitized)
-		if len(m) < 2 {
-			return metricsMode{}, fmt.Errorf("malformed truncate metrics mode: %s", mode)
-		}
-
-		truncLen, err := strconv.Atoi(m[1])
-		if err != nil {
-			return metricsMode{}, fmt.Errorf("malformed truncate metrics mode: %s", mode)
-		}
-
-		if truncLen <= 0 {
-			return metricsMode{}, fmt.Errorf("invalid truncate length: %d", truncLen)
-		}
-
-		return metricsMode{typ: metricModeTruncate, len: truncLen}, nil
-	}
-
-	switch sanitized {
-	case string(metricModeNone):
-		return metricsMode{typ: metricModeNone}, nil
-	case string(metricModeCounts):
-		return metricsMode{typ: metricModeCounts}, nil
-	case string(metricModeFull):
-		return metricsMode{typ: metricModeFull}, nil
-	default:
-		return metricsMode{}, fmt.Errorf("unsupported metrics mode: %s", mode)
-	}
-}
-
-type statisticsCollector struct {
-	fieldID    int
-	icebergTyp iceberg.PrimitiveType
-	mode       metricsMode
-	colName    string
-}
-
 type arrowStatsCollector struct {
 	fieldID     int
 	schema      *iceberg.Schema
@@ -1549,283 +1147,7 @@ func computeStatsPlan(sc *iceberg.Schema, props iceberg.Properties) (map[int]tbl
 	return result, nil
 }
 
-type id2ParquetPath struct {
-	fieldID int
-	path    string
-}
-
-type id2ParquetPathVisitor struct {
-	fieldID int
-	path    []string
-}
-
-func (v *id2ParquetPathVisitor) Schema(_ *iceberg.Schema, res func() []id2ParquetPath) []id2ParquetPath {
-	return res()
-}
-
-func (v *id2ParquetPathVisitor) Struct(_ iceberg.StructType, results []func() []id2ParquetPath) []id2ParquetPath {
-	result := make([]id2ParquetPath, 0, len(results))
-	for _, res := range results {
-		result = append(result, res()...)
-	}
-
-	return result
-}
-
-func (v *id2ParquetPathVisitor) Field(field iceberg.NestedField, res func() []id2ParquetPath) []id2ParquetPath {
-	v.fieldID = field.ID
-	v.path = append(v.path, field.Name)
-	result := res()
-	v.path = v.path[:len(v.path)-1]
-
-	return result
-}
-
-func (v *id2ParquetPathVisitor) List(listType iceberg.ListType, elemResult func() []id2ParquetPath) []id2ParquetPath {
-	v.fieldID = listType.ElementID
-	v.path = append(v.path, "list")
-	result := elemResult()
-	v.path = v.path[:len(v.path)-1]
-
-	return result
-}
-
-func (v *id2ParquetPathVisitor) Map(m iceberg.MapType, keyResult func() []id2ParquetPath, valResult func() []id2ParquetPath) []id2ParquetPath {
-	v.fieldID = m.KeyID
-	v.path = append(v.path, "key_value")
-	keyRes := keyResult()
-	v.path = v.path[:len(v.path)-1]
-
-	v.fieldID = m.ValueID
-	v.path = append(v.path, "key_value")
-	valRes := valResult()
-	v.path = v.path[:len(v.path)-1]
-
-	return append(keyRes, valRes...)
-}
-
-func (v *id2ParquetPathVisitor) Primitive(iceberg.PrimitiveType) []id2ParquetPath {
-	return []id2ParquetPath{{fieldID: v.fieldID, path: strings.Join(v.path, ".")}}
-}
-
-func parquetPathToIDMapping(sc *iceberg.Schema) (map[string]int, error) {
-	result := make(map[string]int)
-
-	paths, err := iceberg.PreOrderVisit(sc, &id2ParquetPathVisitor{})
-	if err != nil {
-		return nil, err
-	}
-
-	for _, entry := range paths {
-		result[entry.path] = entry.fieldID
-	}
-
-	return result, nil
-}
-
-type dataFileStatistics struct {
-	recordCount     int64
-	colSizes        map[int]int64
-	valueCounts     map[int]int64
-	nullValueCounts map[int]int64
-	nanValueCounts  map[int]int64
-	colAggs         map[int]statsAgg
-	splitOffsets    []int64
-}
-
-func (d *dataFileStatistics) partitionValue(field iceberg.PartitionField, sc *iceberg.Schema) any {
-	agg, ok := d.colAggs[field.SourceID]
-	if !ok {
-		return nil
-	}
-
-	if !field.Transform.PreservesOrder() {
-		panic(fmt.Errorf("cannot infer partition value from parquet metadata for a non-linear partition field: %s with transform %s",
-			field.Name, field.Transform))
-	}
-
-	lowerVal, upperVal := agg.min(), agg.max()
-	if !lowerVal.Equals(upperVal) {
-		panic(fmt.Errorf("cannot infer partition value from parquet metadata as there is more than one value for partition field: %s. (low: %s, high: %s)",
-			field.Name, lowerVal, upperVal))
-	}
-
-	val := field.Transform.Apply(must(partitionRecordValue(field, lowerVal, sc)))
-	if !val.Valid {
-		return nil
-	}
-
-	return val.Val.Any()
-}
-
-func (d *dataFileStatistics) toDataFile(schema *iceberg.Schema, spec iceberg.PartitionSpec, path string, format iceberg.FileFormat, filesize int64) iceberg.DataFile {
-	var partitionData map[string]any
-	if !spec.Equals(*iceberg.UnpartitionedSpec) {
-		partitionData = make(map[string]any)
-		for field := range spec.Fields() {
-			val := d.partitionValue(field, schema)
-			if val != nil {
-				partitionData[field.Name] = val
-			}
-		}
-	}
-
-	bldr, err := iceberg.NewDataFileBuilder(spec, iceberg.EntryContentData,
-		path, format, partitionData, d.recordCount, filesize)
-	if err != nil {
-		panic(err)
-	}
-
-	lowerBounds := make(map[int][]byte)
-	upperBounds := make(map[int][]byte)
-
-	for fieldID, agg := range d.colAggs {
-		min := must(agg.minAsBytes())
-		max := must(agg.maxAsBytes())
-		if len(min) > 0 {
-			lowerBounds[fieldID] = min
-		}
-		if len(max) > 0 {
-			upperBounds[fieldID] = max
-		}
-	}
-
-	if len(lowerBounds) > 0 {
-		bldr.LowerBoundValues(lowerBounds)
-	}
-	if len(upperBounds) > 0 {
-		bldr.UpperBoundValues(upperBounds)
-	}
-
-	bldr.ColumnSizes(d.colSizes)
-	bldr.ValueCounts(d.valueCounts)
-	bldr.NullValueCounts(d.nullValueCounts)
-	bldr.NaNValueCounts(d.nanValueCounts)
-	bldr.SplitOffsets(d.splitOffsets)
-
-	return bldr.Build()
-}
-
-func dataFileStatsFromParquetMetadata(pqmeta *metadata.FileMetaData, statsCols map[int]statisticsCollector, colMapping map[string]int) *dataFileStatistics {
-	var (
-		colSizes        = make(map[int]int64)
-		valueCounts     = make(map[int]int64)
-		splitOffsets    = make([]int64, 0)
-		nullValueCounts = make(map[int]int64)
-		nanValueCounts  = make(map[int]int64)
-		invalidateCol   = make(map[int]struct{})
-		colAggs         = make(map[int]statsAgg)
-	)
-
-	for rg := range pqmeta.NumRowGroups() {
-		// reference: https://github.com/apache/iceberg-python/blob/main/pyiceberg/io/pyarrow.py#L2285
-		rowGroup := pqmeta.RowGroup(rg)
-		colChunk, err := rowGroup.ColumnChunk(0)
-		if err != nil {
-			panic(err)
-		}
-
-		dataOffset, dictOffset := colChunk.DataPageOffset(), colChunk.DictionaryPageOffset()
-		if colChunk.HasDictionaryPage() && dictOffset < dataOffset {
-			splitOffsets = append(splitOffsets, dictOffset)
-		} else {
-			splitOffsets = append(splitOffsets, dataOffset)
-		}
-
-		for pos := range rowGroup.NumColumns() {
-			colChunk, err = rowGroup.ColumnChunk(pos)
-			if err != nil {
-				panic(err)
-			}
-
-			fieldID := colMapping[colChunk.PathInSchema().String()]
-			statsCol := statsCols[fieldID]
-			if statsCol.mode.typ == metricModeNone {
-				continue
-			}
-
-			colSizes[fieldID] += colChunk.TotalCompressedSize()
-			valueCounts[fieldID] += colChunk.NumValues()
-			set, err := colChunk.StatsSet()
-			if err != nil {
-				panic(err)
-			}
-
-			if !set {
-				invalidateCol[fieldID] = struct{}{}
-
-				continue
-			}
-
-			stats, err := colChunk.Statistics()
-			if err != nil {
-				invalidateCol[fieldID] = struct{}{}
-
-				continue
-			}
-
-			if stats.HasNullCount() {
-				nullValueCounts[fieldID] += stats.NullCount()
-			}
-
-			if statsCol.mode.typ == metricModeCounts || !stats.HasMinMax() {
-				continue
-			}
-
-			agg, ok := colAggs[fieldID]
-			if !ok {
-				agg, err = createStatsAgg(statsCol.icebergTyp, stats.Type().String(), statsCol.mode.len)
-				if err != nil {
-					panic(err)
-				}
-
-				colAggs[fieldID] = agg
-			}
-
-			switch typ := statsCol.icebergTyp.(type) {
-			case iceberg.BinaryType:
-				stats = wrappedBinaryStats{stats.(*metadata.ByteArrayStatistics)}
-			case iceberg.UUIDType:
-				stats = wrappedUUIDStats{stats.(*metadata.FixedLenByteArrayStatistics)}
-			case iceberg.StringType:
-				stats = wrappedStringStats{stats.(*metadata.ByteArrayStatistics)}
-			case iceberg.FixedType:
-				stats = wrappedFLBAStats{stats.(*metadata.FixedLenByteArrayStatistics)}
-			case iceberg.DecimalType:
-				if typ.Precision() > 18 {
-					stats = wrappedDecStats{stats.(*metadata.FixedLenByteArrayStatistics), typ.Scale()}
-				}
-			}
-
-			agg.update(stats)
-		}
-
-	}
-
-	slices.Sort(splitOffsets)
-	maps.DeleteFunc(nullValueCounts, func(fieldID int, _ int64) bool {
-		_, ok := invalidateCol[fieldID]
-
-		return ok
-	})
-	maps.DeleteFunc(colAggs, func(fieldID int, _ statsAgg) bool {
-		_, ok := invalidateCol[fieldID]
-
-		return ok
-	})
-
-	return &dataFileStatistics{
-		recordCount:     pqmeta.GetNumRows(),
-		colSizes:        colSizes,
-		valueCounts:     valueCounts,
-		nullValueCounts: nullValueCounts,
-		nanValueCounts:  nanValueCounts,
-		splitOffsets:    splitOffsets,
-		colAggs:         colAggs,
-	}
-}
-
-func parquetFilesToDataFiles(fileIO iceio.IO, meta *MetadataBuilder, paths iter.Seq[string]) iter.Seq2[iceberg.DataFile, error] {
+func filesToDataFiles(ctx context.Context, fileIO iceio.IO, meta *MetadataBuilder, paths iter.Seq[string]) iter.Seq2[iceberg.DataFile, error] {
 	return func(yield func(iceberg.DataFile, error) bool) {
 		defer func() {
 			if r := recover(); r != nil {
